pull mapping logic out of collect-documents step into own file
