+++ src/steps/collect-documents.js
@@ -1,5 +1,6 @@
 const ravenClient = require('../raven-client');
 const ravenCache = require('../raven-cache');
+const mapIncludes = require('./map-includes');
 
 const collectDocuments = async (options) => {
   const {
@@ -29,52 +30,12 @@ const collectDocuments = async (options) => {
     documents = await cache.loadDocuments(collection.node);
   }
   else {
-    mapIncludesToDocuments(documents, includes, collection.includes);
+    mapIncludes(documents, includes, collection.includes);
     await cache.saveEtag(collection.node, etag);
     await cache.saveDocuments(collection.node, documents);
   }
   
   return documents;
 };
-
-function mapIncludesToDocuments(documents, includes, collectionIncludes) {
-  if (!collectionIncludes) {
-    return documents;
-  }
-
-  documents.forEach(document => {
-    collectionIncludes.forEach(collectionInclude => {
-      const includeParts = collectionInclude.split(/(\[])?(\.)/).filter(item => item);
-
-      const mapIncludesFromParts = (includeParts, documentPart) => {
-        if (includeParts.length === 1) { // on the final include part
-          if (documentPart && documentPart[includeParts[0]] && includes[documentPart[includeParts[0]]]) { // document part is defined and has a matching include and an include exists for the id in the document
-            const includeId = documentPart[includeParts[0]];
-            documentPart[includeParts[0]] = includes[includeId];
-          }
-          return;
-        }
-
-        switch(includeParts[0]) {
-          case ('[]'):
-            for (let i = 0; i < documentPart.length; i++) {
-              mapIncludesFromParts(includeParts.slice(1), documentPart[i]);
-            }
-            break;
-          case ('.'):
-            mapIncludesFromParts(includeParts.slice(1), documentPart);
-            break;
-          default:
-            mapIncludesFromParts(includeParts.slice(1), documentPart[includeParts[0]]);
-            break;
-        }
-      };
-
-      mapIncludesFromParts(includeParts, document);
-    });
-  });
-
-  return documents;
-}
 
 module.exports = collectDocuments;
